+++ CHANGELOG.md
@@ -18,7 +18,7 @@ SPDX-License-Identifier: GPL-3.0-or-later
 
 ### Added
 
-- Add option `-o` (or `--output-format`) to customize output format
+- Add option `-o` (or `--output-format`) to customize output format, options `-e` and `-m` are now aliases on this option
 - Add official pre-commit hook
 
 ## Version 4.1.0 (2024-10-23)